+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 [Semantic Versioning](http://semver.org/).
 
 ## 1.10.5.dev (development stage/unreleased/unstable)
+### Added
+- fill_up_space_centered()
+### Changed
+- print_stream_info() and print_summary(): unicorn-binance-websocket-api_<version>-python_<version> in top boarder row
 
 ## 1.10.5
 ### Fixed